types: consolidate
